requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0

# Data processing
Pillow>=10.0.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# Optional: For faster training
# albumentations>=1.3.0

# Optional: For experiment tracking
# wandb>=0.15.0
# tensorboard>=2.13.0


